test yuv

+++ dataset.py
@@ -163,7 +163,7 @@ class VideoCaptureYUV:
         elif '.7z' in filename:
             self.height, self.width = 2160,4096
         print(self.width, self.height)
-        self.frame_len = self.width * self.height * 3 / 2
+        self.frame_len = int(self.width * self.height * 3 / 2)
         self.f = open(filename, 'rb')
         self.shape = (int(self.height*1.5), self.width)
